+++ src/cli/utils/vite.js
@@ -11,4 +11,6 @@ export const vite = () => {
 
 // # pnpm
 // pnpm create vite my-vue-app -- --template vue
+
+  // # bun
 }